Rename integration tests' classes
