+++ Kitodo/src/test/java/org/kitodo/services/TaskServiceIT.java
@@ -27,7 +27,7 @@ import static org.junit.Assert.*;
 /**
  * Tests for TaskService class.
  */
-public class TaskServiceTestIT {
+public class TaskServiceIT {
 
     @BeforeClass
     public static void prepareDatabase() throws DAOException {
